put all model methods in db/index
