+++ server/db/index.js
@@ -11,6 +11,71 @@ const {
   CohortUser
 } = models
 
+CohortStretch.getAllCohortStretches = function() {
+  return this.findAll({
+    include: [
+      {
+        model: Cohort,
+        attributes: ['name'],
+        include: [
+          {
+            model: CohortUser,
+            include: [{ model: User, where: { isAdmin: true }, attributes: [] }]
+          }
+        ]
+      }
+    ]
+  }).then(cohortStretches => {
+    return cohortStretches.map(cohortStretch => {
+      const values = cohortStretch.get()
+      const { cohort, ...cohortStretchesFields } = values
+      const cohortValues = cohort.get()
+      const { cohortusers, name } = cohortValues
+      return {
+        ...cohortStretchesFields,
+        cohortName: name,
+        adminIds: cohortusers.map(cu => cu.userId)
+      }
+    })
+  })
+}
+
+Stretch.getAllStretches = function() {
+  return this.findAll({
+    include: [
+      Category,
+      { model: User, as: 'author', attributes: ['firstName', 'lastName'] }
+    ]
+  }).then(stretches => {
+    return stretches.map(stretch => {
+      const data = stretch.get()
+      const { category, author, ...stretchFields } = data
+      return {
+        ...stretchFields,
+        categoryName: category.name,
+        autherName: `${author.firstName} ${author.lastName}`
+      }
+    })
+  })
+}
+
+User.getStudentsByCohort = function(cohortId) {
+  return this.findAll({
+    where: {
+      isAdmin: false
+    },
+    include: [
+      {
+        attributes: [],
+        model: CohortUser,
+        where: {
+          cohortId
+        }
+      }
+    ]
+  })
+}
+
 function initDb(force = false) {
   return db.authenticate().then(() => {
     // Sequelize associations
